CDC improvements: now works with GDC, bugfixes. 64 bit support, various related bugfixes.
Video driver overhaul started:

Draw calls are merged and cached into buffers of vertices stored
in the new GLRenderer class, supporting 3 draw modes - immediate,
vertex arrays and VBOs.
Performance has improved significantly, but more optimization is
possible and monitoring code needs to be rewritten.

All vertex data is now passed to OpenGL as vertex attributes,
not using hardcoded vertices/texcoords/color anymore.
Shaders were changed accordingly.

Added a keyboard shortcut to set draw mode:
1 - Immediate, 2 - Vertex Arrys, 3 - VBOs

+++ formats/pngcommon.d
@@ -306,7 +306,7 @@ struct PNGChunk
     }
 
     ///Get length of the chunk in bytes.
-    uint length(){return data.length + chunk_min_size;}
+    size_t length(){return data.length + chunk_min_size;}
 }
 
 ///Minimum chunk size, taken up by chunk length, type and crc uints.
